+++ packages/mir-website/src/templates/index-page.template.tsx
@@ -1,26 +1,23 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useStyles } from 'react-treat';
+import { IndexPageFrQuery } from '../../types/graphql-types';
 import { BannerPrimary } from '../components/banner-primary';
 import { BlogPreviewSection } from '../components/blog-preview/blog-preview-section';
-import { Footer } from '../components/footer';
 import { InfoSectionType1Group } from '../components/info-section/info-section-type-1-group';
 import { InfoSectionType2 } from '../components/info-section/info-section-type-2';
 import { InfoSectionType3 } from '../components/info-section/info-section-type-3';
 import { InfoSectionType4 } from '../components/info-section/info-section-type-4';
-import { NavBar } from '../components/nav-bar';
 import { Newsletter } from '../components/newsletter/newsletter';
 import { theme } from '../design-system/index';
 import * as stylesRef from './index-page.treat';
-import { PageProps } from '../pages/index.en';
-import { SideBar } from '../components/side-bar';
+
+interface PageProps {
+  data: IndexPageFrQuery;
+  location: Location;
+}
 
 export const Index: React.FC<PageProps> = ({ data, location }) => {
   const styles = useStyles(stylesRef);
-  const [sideMenuState, setSideMenuState] = useState<'openend' | 'closed'>('closed');
-
-  function onOpenSideMenu() {
-    setSideMenuState(sideMenuState === 'openend' ? 'closed' : 'openend');
-  }
 
   const sources = [
     data?.bannerImageMobile?.childImageSharp?.fluid,
@@ -31,16 +28,14 @@ export const Index: React.FC<PageProps> = ({ data, location }) => {
   ];
 
   return (
-    <div className={`${styles.wrapper}`}>
-      <NavBar onOpenSideMenu={onOpenSideMenu}></NavBar>
-      <SideBar className={`${styles.sidebar}`} state={sideMenuState} onOpenSideMenu={onOpenSideMenu}></SideBar>
+    <main className={`${styles.wrapper}`}>
       <BannerPrimary
         imageData={sources}
-        title={data?.gcms?.indexPages[0]?.bannerTitle}
-        subTitle={data?.gcms?.indexPages[0]?.bannerSubTitle}
+        title={data?.gcms?.pages[0]?.bannerTitle}
+        subTitle={data?.gcms?.pages[0]?.bannerSubTitle}
       ></BannerPrimary>
 
-      {data?.gcms?.indexPages[0]?.infoSections.map((section, index) => {
+      {data?.gcms?.pages[0]?.infoSections.map((section, index) => {
         switch (section.type) {
           case 'type1group': {
             return <InfoSectionType1Group key={index} {...section} />;
@@ -61,14 +56,12 @@ export const Index: React.FC<PageProps> = ({ data, location }) => {
       })}
 
       <BlogPreviewSection
-        posts={data?.gcms?.indexPages[0]?.blogSection?.posts}
-        text={data?.gcms?.indexPages[0]?.blogSection?.text}
-        title={data?.gcms?.indexPages[0]?.blogSection?.title}
+        posts={data?.gcms?.pages[0]?.blogSections[0]?.posts}
+        text={data?.gcms?.pages[0]?.blogSections[0]?.text}
+        title={data?.gcms?.pages[0]?.blogSections[0]?.title}
       ></BlogPreviewSection>
 
       <Newsletter id="newsletter"></Newsletter>
-
-      <Footer></Footer>
-    </div>
+    </main>
   );
 };