+++ packages/core-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/newrade/newrade-core/tree/master/packages/core-design-system/compare/core-cli@1.2.0...core-cli@1.2.1) (2020-09-17)
+
+**Note:** Version bump only for package core-cli
+
 # [1.2.0](https://github.com/newrade/newrade-core/tree/master/packages/core-design-system/compare/core-cli@1.1.1...core-cli@1.2.0) (2020-09-16)
 
 ### Bug Fixes